server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.8",
    "google-genai>=1.2.0",
    "google-generativeai>=0.8.4",
    "sqlalchemy>=2.0.38",
]

[tool.uv.workspace]
members = ["server"]
